Create simple_notebook_multiple_inputs.ipynb
This is an adaptation of the simple_notebook jupyter notebook, using multiple inputs.  The multiple inputs are really a singular dictionary that contains entries for each input we want to use.  Because we use a concatenate to bring the two inputs together in an intermediate layer, we must of the functional API, instead of the Sequential model.